refactor bucket
