+++ reductstore/src/storage/bucket.rs
@@ -4,6 +4,7 @@
 mod query;
 mod quotas;
 mod read_only;
+mod remove_entry;
 mod remove_records;
 pub(super) mod settings;
 pub(crate) mod update_records;
@@ -25,7 +26,6 @@ use crate::storage::proto::BucketSettings as ProtoBucketSettings;
 use log::{debug, error};
 use prost::bytes::Bytes;
 use prost::Message;
-use reduct_base::error::ErrorCode;
 use reduct_base::error::ReductError;
 use reduct_base::io::WriteRecord;
 use reduct_base::msg::bucket_api::{BucketInfo, BucketSettings, FullBucketInfo};
@@ -368,92 +368,6 @@ impl Bucket {
         Ok(())
     }
 
-    /// Remove entry from the bucket
-    ///
-    /// # Arguments
-    ///
-    /// * `name` - Entry name.
-    ///
-    /// # Returns
-    ///
-    /// * `HTTPError` - The error if any.
-    pub async fn remove_entry(&self, name: &str) -> Result<(), ReductError> {
-        self.ensure_not_deleting().await?;
-        self.check_mode()?;
-
-        let entry = self.get_entry(name).await?.upgrade()?;
-        entry.mark_deleting().await?;
-
-        let entries = self.entries.clone();
-        let path = self.path.join(name);
-        let bucket_name = self.name.clone();
-        let entry_name = name.to_string();
-        let folder_keeper = self.folder_keeper.clone();
-
-        tokio::spawn(async move {
-            let remove_entry_from_backend = async || {
-                entry.remove_all_blocks().await?;
-                folder_keeper.remove_folder(&entry_name).await?;
-
-                debug!(
-                    "Remove entry '{}' from bucket '{}' and folder '{}'",
-                    entry_name,
-                    bucket_name,
-                    path.display()
-                );
-                entries.write().await?.remove(&entry_name);
-                Ok::<(), ReductError>(())
-            };
-
-            if let Err(err) = remove_entry_from_backend().await {
-                error!(
-                    "Failed to remove entry '{}' from bucket '{}': {}",
-                    entry_name, bucket_name, err
-                );
-            }
-        });
-
-        Ok(())
-    }
-
-    pub(crate) async fn remove_entries_for_bucket_removal(&self) -> Result<(), ReductError> {
-        self.reload().await?;
-
-        let entries_snapshot: Vec<(String, Arc<Entry>)> = {
-            let entries = self.entries.read().await?;
-            entries
-                .iter()
-                .map(|(name, entry)| (name.clone(), Arc::clone(entry)))
-                .collect()
-        };
-
-        for (name, entry) in &entries_snapshot {
-            if let Err(err) = entry.mark_deleting().await {
-                if err.status() != ErrorCode::Conflict {
-                    return Err(err);
-                }
-                if let Ok(status) = entry.status().await {
-                    if status != ResourceStatus::Deleting {
-                        error!(
-                            "Entry '{}' in bucket '{}' is in {:?} state after conflict on delete",
-                            name, self.name, status
-                        );
-                    }
-                }
-            }
-
-            entry.remove_all_blocks().await?;
-            self.folder_keeper.remove_folder(name).await?;
-        }
-
-        let mut entries = self.entries.write().await?;
-        for (name, _) in entries_snapshot {
-            entries.remove(&name);
-        }
-
-        Ok(())
-    }
-
     /// Sync all entries to the file system
     pub async fn sync_fs(&self) -> Result<(), ReductError> {
         if self.cfg.role == InstanceRole::Replica {
@@ -562,37 +476,6 @@ mod tests {
                 bucket.get_or_create_entry("test-1/").await.err(),
                 Some(unprocessable_entity!(
                     "Bucket or entry name can contain only letters, digests and [-,_] symbols"
-                ))
-            );
-        }
-    }
-
-    mod remove_entry {
-        use super::*;
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_remove_entry(#[future] bucket: Arc<Bucket>) {
-            let bucket = bucket.await;
-            write(&bucket, "test-1", 1, b"test").await.unwrap();
-
-            bucket.remove_entry("test-1").await.unwrap();
-            let err = bucket.get_entry("test-1").await.err().unwrap();
-            assert!(
-                err == ReductError::conflict("Entry 'test-1' in bucket 'test' is being deleted")
-                    || err == ReductError::not_found("Entry 'test-1' not found in bucket 'test'"),
-                "Should report deleting or already removed"
-            );
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn test_remove_entry_not_found(#[future] bucket: Arc<Bucket>) {
-            let bucket = bucket.await;
-            assert_eq!(
-                bucket.remove_entry("test-1").await.err(),
-                Some(ReductError::not_found(
-                    "Entry 'test-1' not found in bucket 'test'"
                 ))
             );
         }
@@ -748,24 +631,6 @@ mod tests {
             entry.mark_deleting().await.unwrap();
             assert_eq!(
                 entry.mark_deleting().await,
-                Err(conflict!(
-                    "Entry 'test-1' in bucket 'test' is being deleted"
-                ))
-            );
-        }
-
-        #[rstest]
-        #[tokio::test]
-        async fn remove_entry_returns_conflict_when_entry_is_being_deleted(
-            #[future] bucket: Arc<Bucket>,
-        ) {
-            let bucket = bucket.await;
-            write(&bucket, "test-1", 1, b"test").await.unwrap();
-            let entry = bucket.get_entry("test-1").await.unwrap().upgrade().unwrap();
-            entry.mark_deleting().await.unwrap();
-
-            assert_eq!(
-                bucket.remove_entry("test-1").await,
                 Err(conflict!(
                     "Entry 'test-1' in bucket 'test' is being deleted"
                 ))
